[4337] Move MessagListHeaderView to a different package
