+++ stream-chat-android-ui-components-sample/src/main/res/layout/fragment_component_browser_home.xml
@@ -117,7 +117,7 @@
                 />
 
 
-            <io.getstream.chat.android.ui.feature.messages.list.header.MessageListHeaderView
+            <io.getstream.chat.android.ui.feature.messages.header.MessageListHeaderView
                 android:id="@+id/messagesHeaderView"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
