+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/feature/messages/header/MessageListHeaderView.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.getstream.chat.android.ui.feature.messages.list.header
+package io.getstream.chat.android.ui.feature.messages.header
 
 import android.annotation.SuppressLint
 import android.content.Context
