+++ stream-chat-android-ui-guides/src/main/res/layout/activity_messages.xml
@@ -20,7 +20,7 @@
     android:layout_height="match_parent"
     >
 
-    <io.getstream.chat.android.ui.feature.messages.list.header.MessageListHeaderView
+    <io.getstream.chat.android.ui.feature.messages.header.MessageListHeaderView
         android:id="@+id/messageListHeaderView"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
